+++ domain/study/README.md
@@ -1,0 +1,9 @@
+### domain\Study
+
+| Domain                  | notes                                                                                                                          | googledorks                                      | dig                                         | whoisrws                                                                                               | securitytrails                                                           | dnsdumpster                                              | shodan                                                                                                    | censys                             | affiliation   |
+|:------------------------|:-------------------------------------------------------------------------------------------------------------------------------|:-------------------------------------------------|:--------------------------------------------|:-------------------------------------------------------------------------------------------------------|:-------------------------------------------------------------------------|:---------------------------------------------------------|:----------------------------------------------------------------------------------------------------------|:-----------------------------------|:--------------|
+| securityupdated.co      | relates to APT33 in previous research: https://www.mandiant.com/resources/blog/apt33-insights-into-iranian-cyber-espionage     | securityupdated.com domain still or again active | no relay, domain points to 8.222.152.215    | securityupdated.com -> registered under CN organization but listed for sale                            | two subdomains: hostmaster.securityupdated.com and m.securityupdated.com | two active DNS servers, one in China and one in Germany  | open ports 22, 80, and 8888; several vulnerabilities listed concerning an outdated OpenSSH implementation | no further interesting information | nan           |
+|                         | website became not reachable anymore during the research but site still actively registered                                    |                                                  |                                             | 8.222.152.215 -> located in Singapore, server hosted by Alibaba, 584 other sites hosted on this server |                                                                          | tech used: nginx                                         |                                                                                                           |                                    |               |
+|                         | also checked with tools following this resource which I used in my former job by https://www.bellingcat.com : bit.ly/bcattools |                                                  |                                             |                                                                                                        |                                                                          |                                                          |                                                                                                           |                                    |               |
+| on.tdo.co               | nan                                                                                                                            | nan                                              | on.tdo.com relays to cname.bitly.com        | nothing interesting                                                                                    | nan                                                                      | nginx and WordPress!                                     | ports 80, 443                                                                                             | amazonaws!                         | nan           |
+| on.syn.broadcaster.rock | nan                                                                                                                            | returns nothing on syn.broadcaster.rocks         | syn.broadcaster.rocks relays to 3.141.96.53 | nothing interesting                                                                                    | nan                                                                      | amazonaws!                                               | nothing returned for syn.broadcaster.rocks so used 3.141.96.53; ports 22, 80, 443                         | amazonaws!                         | apt33 (elfin) |
